refactor : GeneratedValue 전략 수정 및 minPrice 타입 int로 통일 (#30)

+++ src/main/java/org/chzz/market/common/validation/validator/ThousandMultipleValidator.java
@@ -4,10 +4,10 @@ import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import org.chzz.market.common.validation.annotation.ThousandMultiple;
 
-public class ThousandMultipleValidator implements ConstraintValidator<ThousandMultiple, Long> {
+public class ThousandMultipleValidator implements ConstraintValidator<ThousandMultiple, Integer> {
 
     @Override
-    public boolean isValid(Long value, ConstraintValidatorContext context) {
+    public boolean isValid(Integer value, ConstraintValidatorContext context) {
         return value != null && value % 1000 == 0 && value > 0;
     }
 }
